+++ infra/env/dev-ecs/locals.tf
@@ -1,17 +1,15 @@
 locals {
-  # Base
-  name = "${var.name_prefix}-${var.environment}" # ex: fastapi-dev
+  name = "${var.name_prefix}-${var.environment}"
 
-  # Logs
   log_group_name = (
     var.log_group_name != "" ? var.log_group_name : "/${var.name_prefix}/${var.environment}"
   )
 
-  # Tags communs
   tags = {
-    Name        = var.name_prefix # ==> "fastapi"
+    Name        = var.name_prefix
+    Project     = var.name_prefix
     ManagedBy   = "Terraform"
-    Environment = var.environment # ==> "dev"
+    Environment = var.environment
   }
 
   kms_account_root        = "arn:${data.aws_partition.current.partition}:iam::${data.aws_caller_identity.current.account_id}:root"